# docs/09-api/openapi.yaml
openapi: 3.0.3
info:
  title: Wovyr AI Platform API
  version: "0.3.0"
  description: |
    The Wovyr single-node server's HTTP API, as actually implemented by
    `crates/wovyr-server` (not the aspirational conventions in `overview.md` —
    see the note below). Source of truth: this spec is generated by hand from
    the Axum route handlers, kept in sync manually until a codegen pipeline
    exists.

    ## Divergence from `overview.md`

    `overview.md` describes a target-state convention (opaque `agt_01H...`
    ids, OAuth2/JWT/mTLS auth, a generic `/operations/{id}` polling resource,
    `RateLimit-*` headers, a uniform `{id, object, tenant, project,
    created_at, updated_at, version}` envelope on every resource). **None of
    that is implemented today.** The real API:

    - Identifies resources by their natural key (agent name, workflow
      `execution_id`, secret `name`, `publisher/name` for plugins/listings) —
      not a generated opaque id.
    - Authenticates via a verified `Authorization: Bearer <credential>`
      (RM-GA-P1 SEC-101, `crates/wovyr-server/src/auth.rs`): `WOVYR_AUTH_MODE=jwt`
      checks a JWT's signature (HS256/RS256) plus expiry/issuer/audience, or
      `apikey` looks up a SHA-256-hashed bearer token — either way the
      verified principal overwrites `X-Wovyr-Principal` before RBAC runs, so a
      client-asserted header alone can no longer win. No OAuth2 authorization
...
      are minted by an external IdP; API keys are minted via
...
servers:
...
components:
  securitySchemes:
    tenantHeader:
      type: apiKey
      in: header
      name: X-Wovyr-Tenant
      description: |
        The acting tenant. Defaults to `"default"` (anonymous/back-compat,
        unrestricted) if absent. Not cryptographically verified — this header
        is a claim, authorized against real memberships server-side.
    principalHeader:
      type: apiKey
      in: header
      name: X-Wovyr-Principal
      description: |
        The acting principal id, used verbatim only in the default
        `disabled-loopback` auth mode. When `bearerAuth` is configured
        (`WOVYR_AUTH_MODE=jwt|apikey`), the server overwrites this with the
        verified principal before any handler runs — a client-supplied value
        is then ignored, not merged.
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        RM-GA-P1 SEC-101. Present only when the server is configured with
        `WOVYR_AUTH_MODE=jwt` (a signed JWT — HS256 via `WOVYR_JWT_HS_SECRET` or
        RS256 via `WOVYR_JWT_RS_PUBLIC_KEY`, `sub` claim becomes the principal)
        or `WOVYR_AUTH_MODE=apikey` (a bearer token minted via
        `wovyr auth create-key`, SHA-256-hashed server-side). The default
        `disabled-loopback` mode carries no bearer credential at all and
        instead trusts `X-Wovyr-Principal`/`X-Wovyr-Tenant` verbatim — a
        local-dev fallback the server refuses to expose on a non-loopback
        bind unless `WOVYR_ALLOW_ANONYMOUS=1` is set.
  parameters:
    limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    cursor:
      name: cursor
      in: query
      schema: { type: string }
      description: Opaque pagination cursor from a previous response's `next_cursor`.
    xWovyrTenant:
      name: X-Wovyr-Tenant
      in: header
      schema: { type: string, default: default }
    xWovyrPrincipal:
      name: X-Wovyr-Principal
      in: header
      schema: { type: string }
...
    idempotencyKey:
      name: Idempotency-Key
      in: header
      schema: { type: string }
      description: >-
        RM-GA-P4 API-703: a repeated key on this route replays the original
        (tenant, method, path, key)-scoped response instead of re-executing
        the mutation. Only a successful response is cached.
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, type, status, request_id]
          properties:
            code: { type: string, example: validation_failed }
            message: { type: string }
            type: { type: string, enum: [client_error, server_error] }
            status: { type: integer }
            request_id: { type: string }
    PageMeta:
      type: object
      properties:
        has_more: { type: boolean }
        next_cursor: { type: string, nullable: true }
        total_estimate: { type: integer }
...
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
paths:
...
  /api/v1/secrets:
    get:
      tags: [secrets]
      summary: List secret metadata (never values) for the caller's tenant
      parameters:
        - $ref: '#/components/parameters/xWovyrTenant'
        - $ref: '#/components/parameters/xWovyrPrincipal'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/cursor'
      responses:
        '200':
          description: Paginated secret metadata
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageMeta'
                  - type: object
                    properties: { data: { type: array, items: { type: object } } }
    post:
      tags: [secrets]
      summary: Create a secret
      parameters:
        - $ref: '#/components/parameters/xWovyrTenant'
        - $ref: '#/components/parameters/xWovyrPrincipal'
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, value]
              properties: { name: { type: string }, value: { type: string } }
      responses:
        '200':
          description: Created (metadata only — value never returned)
          content:
            application/json: { schema: { type: object } }
  /api/v1/secrets/{name}:
    get:
      tags: [secrets]
      summary: Get one secret's metadata
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/xWovyrTenant'
        - $ref: '#/components/parameters/xWovyrPrincipal'
      responses:
        '200': { description: Metadata }
        '404': { $ref: '#/components/responses/Error' }
    delete:
      tags: [secrets]
      summary: Delete a secret
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/xWovyrTenant'
        - $ref: '#/components/parameters/xWovyrPrincipal'
        - $ref: '#/components/parameters/idempotencyKey'
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/Error' }
  /api/v1/secrets/{name}/rotate:
    post:
      tags: [secrets]
      summary: Rotate a secret's value (retains one previous version)
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/xWovyrTenant'
        - $ref: '#/components/parameters/xWovyrPrincipal'
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [value], properties: { value: { type: string } } }
      responses:
        '200': { description: Rotated (metadata only) }

  # -------------------------------------------------------------- tenancy --
...
security:
  # Alternatives (a client uses whichever the server is actually configured
  # for — `WOVYR_AUTH_MODE`): the local-dev default (raw, unverified headers),
  # or a verified bearer credential (RM-GA-P1 SEC-101) alongside the tenant
  # header (still asserted, but authorized against real memberships).
  - tenantHeader: []
    principalHeader: []
  - tenantHeader: []
    bearerAuth: []